+++ mongodb/src/test/scala/loco/repository/ITTest.scala
@@ -8,4 +8,8 @@ trait ITTest extends Suite with BeforeAndAfterEach with BeforeAndAfterAll {
   override def beforeAll(): Unit = {
     EmbeddedDBEnv.givenUniqueIndex()
   }
+
+  override def afterAll(): Unit = {
+    EmbeddedDBEnv.process.stop()
+  }
 }
